refactor: reorganize feature flag system into features folder structure
- Move all feature flag files to src/features/feature-flags/
- Organize into logical subdirectories: types, lib, contexts, hooks, components
- Create index.ts for clean exports
- Update all import paths to use new structure
- Remove old scattered files
- Update documentation to reflect new organization
