+++ src/features/feature-flags/index.ts
@@ -1,0 +1,18 @@
+// Feature Flag System - Main Export
+
+// Types
+export * from './types';
+
+// Configuration
+export * from './lib/config';
+
+// Context and Hooks
+export { FeatureFlagProvider, useFeatureFlags } from './contexts/FeatureFlagContext';
+export { useFeatureFlag } from './hooks/useFeatureFlag';
+
+// Components
+export { FeatureFlag } from './components/FeatureFlag';
+export { FeatureFlagTest } from './components/FeatureFlagTest';
+
+// Constants
+export { FEATURE_FLAGS, USER_TYPES, USER_GROUPS } from './lib/config';
